backend/requirements.txt
flask
flask-cors
torch
torchvision
diffusers
transformers
peft
accelerate
tqdm
Pillow
safetensors
scipy
huggingface_hub
